+++ CHANGELOG.md
@@ -12,6 +12,12 @@ These changes are available on the `master` branch, but have not yet been releas
 
 ### Added
 
+- Added `ScheduledEventEntityType` enum and `ScheduledEventEntityMetadata` class, added
+  missing attributes and API-aligned parameters to `ScheduledEvent`,
+  `Guild.create_scheduled_event`, and `ScheduledEvent.edit`, and added `use_cache` to
+  `ScheduledEvent.subscribers()`.
+  ([#3025](https://github.com/Pycord-Development/pycord/pull/3025))
+
 ### Changed
 
 ### Fixed
@@ -22,8 +28,14 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#3231](https://github.com/Pycord-Development/pycord/pull/3231))
 - Allow `ForumTag` to be created without an emoji.
   ([#3245](https://github.com/Pycord-Development/pycord/pull/3245))
+- Fixed `ScheduledEvent` subscriber cache not being kept in sync correctly.
+  ([#3025](https://github.com/Pycord-Development/pycord/pull/3025))
 
 ### Deprecated
+
+- Deprecated `ScheduledEventLocationType`, `ScheduledEventLocation`, and several
+  `ScheduledEvent` attributes and related methods' parameters in favor of their
+  API-aligned names. ([#3025](https://github.com/Pycord-Development/pycord/pull/3025))
 
 ### Removed